requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
openai
sentence-transformers
sqlalchemy
pydantic
rich
